requirements.txt
torch
torchvision
transformers>=4.46.0
gradio
modelscope
pyyaml